Add GenerateMigration project
GenerateMigration project will load the schemas of a before and after
database, generate a migration class with up and down to migrate the old
to the new and add that class file to a pre-existing project at a given
path.

The text template used to generate the code isn't all that clean, but it
works with adding and removing tables, columns from tables, and foreign
keys, tested on SQL server.