Set default values when adding columns to a table
Add code to the default generator and the template so that, when adding new
columns to a table, if the column is non-nullable, set a default value for
it so that the migration will still work if there is data in the table.
Also improve the spacing of the foreign keys in the output code.